refactor(http-server): split executor tests
